backend/requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
pandas
numpy
matplotlib
seaborn
joblib
imbalanced-learn
python-dotenv
psycopg2-binary
geopy
googlemaps